# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/experiment": {
            "put": {
                "description": "Update configuration db with experiment data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "experiments-params"
                ],
                "summary": "updates experiment data and return experiment",
                "parameters": [
                    {
                        "description": "Experiment data",
                        "name": "experiment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Experiment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create experiment in configuration db.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "experiments-params"
                ],
                "summary": "creates experiment data and return experiment",
                "parameters": [
                    {
                        "description": "Experiment data",
                        "name": "experiment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Experiment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/experiment/parameter": {
            "post": {
                "description": "add or create experiment/arm parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "experiments-params"
                ],
                "summary": "add a parameter to an experiment/arm or update the parameter value if already exists.",
                "parameters": [
                    {
                        "description": "Reward data parameter",
                        "name": "experiment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RewardDataParameter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/experiment/{experiment_id}": {
            "get": {
                "description": "Consults configuration DB to fetch experiment data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "experiments-params"
                ],
                "summary": "return experiment data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Experiment ID",
                        "name": "experiment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/experiments": {
            "get": {
                "description": "Consults configuration DB to fetch experiments data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "experiments-params"
                ],
                "summary": "return data from all experiments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Arm": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Experiment": {
            "type": "object",
            "properties": {
                "arms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Arm"
                    }
                },
                "experiment_id": {
                    "type": "string"
                },
                "parameters": {},
                "policy_type": {
                    "type": "string"
                }
            }
        },
        "model.RewardDataParameter": {
            "type": "object",
            "properties": {
                "arm": {
                    "type": "string"
                },
                "experiment_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "value": {}
            }
        }
    }
}